postgre

+++ app/Controllers/Pelaku.php
@@ -7,6 +7,7 @@ use App\Models\DetailJenisNarkobaModel;
 use App\Models\JenisNarkobaModel;
 use App\Models\KasusModel;
 use App\Models\PelakuViewModel;
+use App\Models\PenangananModel;
 use Config\Services;
 
 class Pelaku extends BaseController
@@ -21,17 +22,12 @@ class Pelaku extends BaseController
 
     public function index()
     {
-        
-        // $data['pelakus'] = $pelaku->getWhere(['id_pelaku' => '1'])->getResultArray();
         $pelaku = new PelakuViewModel();
         $data['pelakus'] = $pelaku->findAll();
         $detail = new DetailJenisNarkobaModel();
         $data['details'] = $detail->findAll();
         $kasus = new KasusModel();
         $data['kasuss'] = $kasus->findAll();
-        // isset($data['pelakus']) ? count($data['pelakus']) : 0;
-
-        
         
 		echo view('pelaku', $data);
     }
@@ -44,8 +40,10 @@ class Pelaku extends BaseController
         $kasus = new KasusModel();
         $data['kasuss'] = $kasus->findAll();
         $pelaku = new PelakuViewModel();
-        $data['kurir'] = $pelaku->query('Select * from v_pelaku where profil="Kurir"' )->getResultArray();
-        $data['bandar'] = $pelaku->query('Select * from v_pelaku where profil="Bandar" ' )->getResultArray();
+        $data['kurir'] = $pelaku->query("Select * from v_pelaku where profil='Kurir'" )->getResultArray();
+        $data['bandar'] = $pelaku->query("Select * from v_pelaku where profil='Bandar' " )->getResultArray();
+        $penanganan = new PenangananModel();
+        $data['penanganan'] = $penanganan->findAll();
         echo view('form_pelaku', $data);
     }
     
@@ -57,7 +55,6 @@ class Pelaku extends BaseController
             'nama' => 'required',
             'jenis_kelamin' => 'required',
             'profil' => 'required',
-            // 'file_upload' => 'uploaded[file_upload]|mime_in[file_upload,image/jpg,image/jpeg,image/gif,image/png]|max_size[file_upload,4096]'
         ]);
         $isDataValid = $validation->withRequest($this->request)->run();
 
@@ -83,15 +80,12 @@ class Pelaku extends BaseController
                     $foto="Wanita.jpg";
                 }
             }else{
-                $nama=$this->request->getPost('nama').".jpg";
+                $nama=mt_rand(100000, 999999)."-".$this->request->getPost('nama').".jpg";
                 $path = WRITEPATH . '../public/assets/images/foto/';
                 $upload->move($path,$nama);
                 $foto = $nama;
             }
             
-            
-            // $dataBerkas = $this->request->getFile('foto');
-            // $fileName = $dataBerkas->getName();
             $pelaku->insert([
                 "nama" => $this->request->getPost('nama'),
                 "tempat_lahir" => $this->request->getPost('tempat_lahir'),
@@ -99,7 +93,6 @@ class Pelaku extends BaseController
                 "nik" => $this->request->getPost('nik'),
                 "profil" => $this->request->getPost('profil'),
                 "jenis_kelamin" => $this->request->getPost('jenis_kelamin'),
-                // "jenis_narkoba" => array_implode($j),
                 "berat" => $this->request->getPost('berat'),
                 "uang_sita" => $this->request->getPost('uang_sita'),
                 "tkp" => $this->request->getPost('tkp'),
@@ -108,25 +101,32 @@ class Pelaku extends BaseController
                 "id_kasus" => $this->request->getPost('id_kasus'),
                 "id_atasan" => $idatasan,
                 'foto' => $foto,
-                // 'foto' => $fileName
+                "id_penanganan" => $this->request->getPost('id_penanganan'),
+                "tempat_penanganan" => $this->request->getPost('tempat_penanganan'),
+                "tanggal_penangkapan" => $this->request->getPost('tanggal_penangkapan'),
             ]);
-            // $dataBerkas->move(WRITEPATH . '../public/assets/images/');
             $id_pelaku = $pelaku->getInsertID();
-            $jenis_narkoba = count($this->request->getPost('jenis_narkoba'));
-            for ($i = 0; $i < $jenis_narkoba; $i++) {
-                $datas[$i] = array(
-                    'id_pelaku' => $id_pelaku,
-                    'id_jenis_narkoba' => $this->request->getPost('jenis_narkoba[' . $i . ']'),
-                );
-                $jenis = new DetailJenisNarkobaModel();
-                $jenis->insert($datas[$i]);
-            }
+            $this->insertNarkoba($id_pelaku);
             $this->session->setFlashdata('success', 'Data Berhasil Ditambah');
             
         }else{
             $this->session->setFlashdata('error', $validation->getError());
         }
         return redirect('admin/pelaku');
+        
+    }
+
+    public function insertNarkoba($id)
+    {
+        $jenis_narkoba = count($this->request->getPost('jenis_narkoba'));
+            for ($i = 0; $i < $jenis_narkoba; $i++) {
+                $datas[$i] = array(
+                    'id_pelaku' => $id,
+                    'id_jenis_narkoba' => $this->request->getPost('jenis_narkoba[' . $i . ']'),
+                );
+                $jenis = new DetailJenisNarkobaModel();
+                $jenis->insert($datas[$i]);
+            }
     }
 
     public function get_atasan(){
@@ -154,6 +154,8 @@ class Pelaku extends BaseController
         $data['jeniss'] = $jenis->findAll();
         $kasus = new KasusModel();
         $data['kasuss'] = $kasus->findAll();
+        $penanganan = new PenangananModel();
+        $data['penanganan'] = $penanganan->findAll();
         echo view('form_pelaku_edit', $data);
     }
 
@@ -172,18 +174,21 @@ class Pelaku extends BaseController
             $idatasan= $this->request->getPost('id_atasan');
         }
         if($isDataValid){
+            $fotolama = $this->request->getPost('file_upload_old');
             $upload = $this->request->getFile('file_upload');
-            if($upload->getName()==""){
-                $foto=$this->request->getPost('file_upload_old');
+            $uploadName = $upload->getName();
+            if($uploadName==""){
+                $foto=$fotolama;
             }else{
-                // if($upload != 'Pria.jpg' || $upload != 'Wanita.jpg'){
-                //     unlink("../public/assets/images/foto/".$upload);
-                // }
-                $nama=$this->request->getPost('nama').".jpg";
+                if($uploadName != 'Pria.jpg' AND $uploadName != 'Wanita.jpg'){
+                    $delete = unlink( $_SERVER['DOCUMENT_ROOT']."/assets/images/foto/$fotolama");
+                }
+                $nama=mt_rand(100000, 999999)."-".$this->request->getPost('nama').'.jpg';
                 $path = WRITEPATH . '../public/assets/images/foto/';
                 $upload->move($path,$nama);
                 $foto = $nama;
             }
+            
             $pelaku->update($id, [
                 "nama" => $this->request->getPost('nama'),
                 "tempat_lahir" => $this->request->getPost('tempat_lahir'),
@@ -191,7 +196,6 @@ class Pelaku extends BaseController
                 "nik" => $this->request->getPost('nik'),
                 "profil" => $this->request->getPost('profil'),
                 "jenis_kelamin" => $this->request->getPost('jenis_kelamin'),
-                // "jenis_narkoba" => array_implode($j),
                 "berat" => $this->request->getPost('berat'),
                 "uang_sita" => $this->request->getPost('uang_sita'),
                 "tkp" => $this->request->getPost('tkp'),
@@ -199,31 +203,42 @@ class Pelaku extends BaseController
                 "unit" => $this->request->getPost('unit'),
                 "id_kasus" => $this->request->getPost('id_kasus'),
                 "id_atasan" => $idatasan,
-                'foto' => $foto
+                'foto' => $foto,
+                "id_penanganan" => $this->request->getPost('id_penanganan'),
+                "tempat_penanganan" => $this->request->getPost('tempat_penanganan'),
+                "tanggal_penangkapan" => $this->request->getPost('tanggal_penangkapan'),
             ]);
             
             $jenis = new DetailJenisNarkobaModel();
             $jenis->delete($id);
 
             $id_pelaku = $id;
-            $jenis_narkoba = count($this->request->getPost('jenis_narkoba'));
-            for ($i = 0; $i < $jenis_narkoba; $i++) {
-                $datas[$i] = array(
-                    'id_pelaku' => $id_pelaku,
-                    'id_jenis_narkoba' => $this->request->getPost('jenis_narkoba[' . $i . ']')
-                );
-                $jenis = new DetailJenisNarkobaModel();
-                $jenis->insert($datas[$i]);
-            }
+            $this->insertNarkoba($id_pelaku);
+            
             session()->setFlashdata('success', 'Berkas Berhasil Diupdate');
             return redirect('admin/pelaku');
         }
     }
 
     public function deletePelaku($id){
+        
+       
         $pelaku = new PelakuModel();
-        $pelaku->delete($id);
-        session()->setFlashdata('success', 'Berkas Berhasil Dihapus');
+        $data['pelaku'] = $pelaku->where('id_pelaku', $id)->first();
+        $foto = $data['pelaku']['foto'];
+        if($foto != 'Pria.jpg' AND $foto != 'Wanita.jpg'){
+            $delete = unlink( $_SERVER['DOCUMENT_ROOT']."/assets/images/foto/$foto");
+        }else{
+            $delete=true;
+        }
+        
+        if($delete){
+            $pelaku->delete($id);
+            session()->setFlashdata('success', 'Berkas Berhasil Dihapus');
+        }else{
+            session()->setFlashdata('error', 'Berkas Gagal Dihapus');
+        }
+        
         return redirect('admin/pelaku');
     }
 }
